Add 'include' and 'exclude' lists
Allow the user to specify a list of tracker IDs to explicitly include
or exclude. Using the 'include' list can be used to mimic the
behaviour of the mobile phone application to only sync the owner's
tracker. The 'exclude' list can be used, for example, to avoid syncing
a tracker with a known problem.